flaud-book/pyproject.toml
[tool.poetry]
name = "flaud-book"
version = "0.1.0"
description = ""
authors = ["dotpep <pepelnine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
sqlalchemy-utils = "^0.41.2"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
